requirements.txt
streamlit>=1.28.0
pytesseract>=0.3.10
Pillow>=10.0.0
pandas>=2.0.0
opencv-python>=4.8.0
python-dateutil>=2.8.2
plotly>=5.15.0